backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<3.16
djangorestframework-simplejwt>=5.3,<5.4
django-cors-headers>=4.4,<4.5
psycopg2-binary>=2.9,<2.10
redis>=5.0,<5.1
celery>=5.4,<5.5
channels>=4.1,<4.2
channels-redis>=4.2,<4.3
